delete profile page added

+++ server/controllers/userControllers.js
@@ -80,10 +80,21 @@ export const loggedUser = async (req, res) => {
 
 //Delete profile
 export const deleteUser = async (req, res) => {
-  const { userId } = req.params;
+  const { userId, password } = req.body;
+
+  console.log("body==>", req.body);
 
   try {
     const findUser = await User.findById(userId);
+
+    const isMatched = await bcrypt.compare(password, findUser.password);
+
+    if (!isMatched || !findUser)
+      return res.status(400).send({
+        success: false,
+        error: "Please type the correct password",
+      });
+
     if (findUser.image) {
       const filePath = "uploads/profileImage/" + findUser.image;
       fs.unlink(filePath, (err) => {
@@ -94,6 +105,8 @@ export const deleteUser = async (req, res) => {
         console.log("File deleted successfully");
       });
     }
+
+    console.log("User matched==>", isMatched);
 
     const deletedUser = await User.findByIdAndDelete(userId);
